Restructuring of files, improved size handling, scrolling and logo
