+++ react/src/lib/components/Menu/components/FilterInput/index.ts
@@ -1,0 +1,1 @@
+export { FilterInput } from "./FilterInput";
