+++ react/src/lib/components/Menu/components/Group/Group.tsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { Icon as EdsIcon } from "@equinor/eds-core-react";
 import { arrow_drop_right, arrow_drop_down } from "@equinor/eds-icons";
 
-import { Icon } from "./Icon";
+import { Icon } from "../Icon";
 
 EdsIcon.add({ arrow_drop_down, arrow_drop_right });
 
@@ -12,11 +12,12 @@ import "./Group.css";
 type GroupProps = {
     title: string;
     icon?: string;
+    open?: boolean;
     children?: React.ReactNode;
 };
 
 export const Group: React.FC<GroupProps> = (props) => {
-    const [collapsed, setCollapsed] = React.useState<boolean>(true);
+    const [collapsed, setCollapsed] = React.useState<boolean>(!props.open);
     return (
         <div className="Group">
             <div
@@ -48,6 +49,7 @@ export const Group: React.FC<GroupProps> = (props) => {
 Group.propTypes = {
     title: PropTypes.string.isRequired,
     icon: PropTypes.string,
+    open: PropTypes.bool,
     children: PropTypes.oneOfType([
         PropTypes.arrayOf(PropTypes.node),
         PropTypes.node,
