+++ react/src/lib/components/Menu/components/Group/index.ts
@@ -1,0 +1,1 @@
+export { Group } from "./Group";
