+++ react/src/lib/components/Menu/components/Icon/Icon.tsx
