--- react/src/lib/components/Menu/components/Logo.css
@@ -1,15 +1,0 @@
-.Menu .Logo {
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: 60%;
-}
-
-.Menu .LogoLarge {
-    width: 100%;
-    height: 160px;
-}
-
-.Menu .LogoSmall {
-    width: 48px;
-    height: 48px;
-}
