+++ react/src/lib/components/Menu/components/MenuBar/index.ts
@@ -1,0 +1,1 @@
+export { MenuBar } from "./MenuBar";
