+++ react/src/lib/components/Menu/components/MenuContent/index.ts
@@ -1,0 +1,1 @@
+export { MenuContent } from "./MenuContent";
