+++ react/src/lib/components/Menu/components/MenuDrawer/index.ts
@@ -1,0 +1,1 @@
+export { MenuDrawer } from "./MenuDrawer";
