+++ react/src/lib/components/Menu/components/Page/index.ts
@@ -1,0 +1,1 @@
+export { Page } from "./Page";
