+++ react/src/lib/components/Menu/components/ScrollArea/ScrollArea.css
@@ -1,0 +1,28 @@
+.Menu .ScrollArea {
+    overflow: hidden;
+    position: relative;
+    width: 100%;
+    display: block;
+    flex-grow: 1;
+}
+
+.Menu .ScrollArea .Content {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+}
+
+.Menu .VerticalScrollBar {
+    position: absolute;
+    right: 8px;
+    width: 8px;
+    min-height: 80px;
+    background-color: #5c5c5c;
+    border-radius: 4px;
+    z-index: 10;
+}
+
+.Menu .VerticalScrollBar:hover {
+    opacity: 1;
+}
