+++ react/src/lib/components/Menu/components/ScrollArea/index.ts
@@ -1,0 +1,1 @@
+export { ScrollArea } from "./ScrollArea";
