+++ react/src/lib/components/Menu/components/Section/Section.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { Icon } from "./Icon";
+import { Icon } from "../Icon";
 
 import "./Section.css";
