+++ react/src/lib/components/Menu/components/Section/index.ts
@@ -1,0 +1,1 @@
+export { Section } from "./Section";
