+++ react/src/lib/components/Menu/components/TopMenu/index.ts
@@ -1,0 +1,1 @@
+export { TopMenu } from "./TopMenu";
