+++ react/src/lib/components/Menu/hooks/usePrevious.ts
@@ -1,0 +1,11 @@
+import { useRef, useEffect } from "react";
+
+export const usePrevious = <T>(value: T): T | undefined => {
+    const ref = useRef<T>();
+
+    useEffect(() => {
+        ref.current = value;
+    });
+
+    return ref.current;
+};
